[pobj] laser added; we can place mirrors

+++ inc/player.h
@@ -15,7 +15,7 @@ enum PlayerEmote {
 
 
 void PLR_init();
-bool PLR_goto(f16 dest_x, f16 dest_y);
+bool PLR_goto(u16 dest_x, u16 dest_y); // in integer tile coords
 u16 PLR_curr_tile_x();
 u16 PLR_curr_tile_y();
 void PLR_update(u32 framecounter);
